langchain-python-service/requirements.txt
# Core framework
fastapi>=0.115.14
uvicorn[standard]>=0.35.0
pydantic>=2.11.7

# LangChain - Latest versions
langchain>=0.3.26
langchain-openai>=0.3.27
langchain-core>=0.3.66
langchain-community>=0.3.26

# HTTP client
httpx>=0.28.1
aiohttp>=3.12.13

# Environment
python-dotenv>=1.1.1

# Multiprocessing
multiprocess>=0.70.18

# Utilities
pydantic-settings>=2.10.1
loguru>=0.7.3 